# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: 뭉치장 API
  description: |
    유행 음식 공동구매 플랫폼 뭉치장의 API 명세입니다.

    ## 인증
    로그인이 필요한 API는 `Authorization: Bearer {accessToken}` 헤더를 포함해야 합니다.

    ## 공통 응답 형식
    모든 응답은 아래 구조를 따릅니다.
    ```json
    { "success": true, "data": { ... }, "error": null }
    { "success": false, "data": null, "error": { "code": "ERROR_CODE", "message": "설명", "detail": "상세" } }
    ```
  version: v1
servers:
  - url: https://api.moongchijang.com
    description: 운영 서버
  - url: https://api.moongchijang.com/dev
    description: 개발 서버
  - url: http://localhost:8080
    description: 로컬 서버
...
paths:
...
  /api/v1/users/me/participations:
    get:
      tags: [MyPage]
      summary: 내 참여 내역 상태별 조회
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [IN_PROGRESS, PICKUP_WAITING, PICKUP_COMPLETED, CANCELLED_OR_REFUNDED]
          description: IN_PROGRESS=진행 중 / PICKUP_WAITING=픽업 대기 / PICKUP_COMPLETED=픽업 완료 / CANCELLED_OR_REFUNDED=환불·취소
      responses:
        '200':
          description: 상태별 참여 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMypageParticipationList'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/users/me/participations/in-progress:
    get:
      tags: [MyPage]
      summary: 진행 중 탭 참여 공구 목록 조회
      description: 페이지네이션 기반으로 진행 중 탭 카드를 조회한다. 정렬 기준은 참여일시(participatedAt) 내림차순이다.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 진행 중 참여 공구 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInProgressParticipationPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/users/me/participations/pickup-waiting:
    get:
      tags: [MyPage]
      summary: 픽업 대기 탭 참여 완료 공구 이력 조회
      description: 페이지네이션 기반으로 픽업 대기 탭 카드를 조회한다. 정렬 기준은 참여일시(participatedAt) 내림차순이다.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 픽업 대기 참여 공구 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePickupWaitingParticipationPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/images/presigned-urls:
    post:
      tags: [ImageUpload]
      summary: S3 이미지 업로드용 Presigned URL 발급
      description: |
        인증된 사용자의 현재 역할이 BUYER, SELLER, ADMIN 중 하나이면 호출할 수 있다.
        썸네일(최대 1장), 상품 이미지(최대 10장)에 대한 업로드 URL을 발급한다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImagePresignedUploadRequest'
      responses:
        '200':
          description: Presigned URL 발급 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseImagePresignedUpload'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/images:
    delete:
      tags: [ImageUpload]
      summary: S3 업로드 이미지 삭제
      description: |
        인증된 사용자의 현재 역할이 BUYER, SELLER, ADMIN 중 하나이면 호출할 수 있다.
        요청한 S3 key 목록을 삭제한다. 허용된 prefix 범위의 key만 삭제 가능하다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageDeleteRequest'
      responses:
        '200':
          description: 이미지 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseImageDelete'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

# ─────────────────────────────────────────────────────────────────
# Components
# ─────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
  responses:
...
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
...
  schemas:
...
    ApiResponseError:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: false }
        data: { nullable: true, example: null }
        error:
          type: object
          required: [code, message, detail]
          properties:
            code: { type: string, example: GROUPBUY_NOT_FOUND }
            message: { type: string, example: 존재하지 않는 공구입니다 }
            detail: { type: string, nullable: true }

    # ── Auth ───────────────────────────────────────────────────
...
    ApiResponseMypageParticipationList:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            type: object
            required: [participationId, groupBuyId, thumbnailUrl, productName, participationStatus, achievementRate, achievementStatus, displayStatus, storeName, pickupDate, pickupTimeStart, pickupTimeEnd, pickupLocation, paymentAmount, paidAt, paymentMethod, quantity, pickupStatus, dDay, canCancel, canViewPickup, canViewQr, qrAvailability]
            properties:
              participationId: { type: integer, format: int64 }
              groupBuyId: { type: integer, format: int64 }
              thumbnailUrl: { type: string, nullable: true, description: 카드 상품 이미지 URL }
              productName: { type: string }
              participationStatus:
                type: string
                enum: [PENDING, PAID_WAITING_GOAL, CONFIRMED, CANCELLED, REFUND_PENDING, REFUNDED]
              achievementRate:
                type: integer
                minimum: 0
                maximum: 100
                description: 현재 수량 기준 달성률
              achievementStatus:
                type: string
                enum: [BEFORE_ACHIEVED, ACHIEVED]
              displayStatus:
                type: string
                enum: [PICKED_UP, PAID_WAITING_GOAL, CONFIRMED, REFUND_PENDING, REFUNDED, PENDING, CANCELLED]
                description: 픽업/참여 상태 조합 기반 화면 표시 상태
              storeName: { type: string }
              pickupDate: { type: string, format: date }
              pickupTimeStart: { type: string, format: time }
              pickupTimeEnd: { type: string, format: time }
              pickupLocation: { type: string }
              paymentAmount: { type: integer }
              paidAt: { type: string, format: date-time, nullable: true, description: 주문 상세 결제 날짜 }
              paymentMethod: { type: string, nullable: true, description: 결제 내역 결제 수단 }
              quantity: { type: integer }
              pickupStatus:
                type: string
                enum: [NOT_READY, READY, PICKED_UP, NO_SHOW]
              dDay:
                type: integer
                description: 공구 마감일까지 남은 일수
              canCancel:
                type: boolean
                description: PAID_WAITING_GOAL 참여이고 APPROVED 결제 주문이 있으면 true
              canViewPickup: { type: boolean, description: "참여 확정 후 픽업 미완료이면 /api/v1/participations/{participationId}/pickup 호출 가능" }
              canViewQr: { type: boolean, description: "참여 확정 후 픽업 미완료이면 /api/v1/participations/{participationId}/qr 호출 가능. 픽업일 전에는 QR API가 LOCKED를 반환" }
              qrAvailability:
                type: string
                enum: [UNAVAILABLE, LOCKED, AVAILABLE, PICKED_UP]
        error: { nullable: true, example: null }
...
    ApiResponseInProgressParticipationPage:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          required: [content, totalElements, totalPages]
          properties:
            content:
              type: array
              items:
                type: object
                required: [participationId, groupBuyId, productName, storeName, pickupAt, paidAmount, quantity, achievementRate, dDay, participatedAt]
                properties:
                  participationId: { type: integer, format: int64 }
                  groupBuyId: { type: integer, format: int64 }
                  productName: { type: string }
                  storeName: { type: string }
                  pickupAt: { type: string, format: date-time }
                  paidAmount: { type: integer }
                  quantity: { type: integer }
                  achievementRate: { type: integer, description: 0~100 정수 퍼센트 }
                  dDay: { type: integer, description: 마감일까지 남은 일수 }
                  participatedAt: { type: string, format: date-time }
            totalElements: { type: integer }
            totalPages: { type: integer }
        error: { nullable: true, example: null }
    ApiResponsePickupWaitingParticipationPage:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          required: [content, totalElements, totalPages]
          properties:
            content:
              type: array
              items:
                type: object
                required: [participationId, groupBuyId, productName, storeName, pickupAt, paidAmount, quantity, isClosed, participatedAt]
                properties:
                  participationId: { type: integer, format: int64 }
                  groupBuyId: { type: integer, format: int64 }
                  productName: { type: string }
                  storeName: { type: string }
                  pickupAt: { type: string, format: date-time }
                  paidAmount: { type: integer }
                  quantity: { type: integer }
                  isClosed: { type: boolean, description: 마감 공구 여부 }
                  participatedAt: { type: string, format: date-time }
            totalElements: { type: integer }
            totalPages: { type: integer }
        error: { nullable: true, example: null }
...
    ImageUploadCategory:
      type: string
      description: 이미지 업로드 카테고리
      enum: [THUMBNAIL, PRODUCT]
    ImagePresignedUploadItemRequest:
      type: object
      required: [category, fileName, contentType]
      properties:
        category:
          $ref: '#/components/schemas/ImageUploadCategory'
        fileName:
          type: string
          description: 원본 파일명
          example: cake.jpg
        contentType:
          type: string
          description: 파일 Content-Type
          example: image/jpeg
    ImagePresignedUploadRequest:
      type: object
      required: [files]
      properties:
        groupBuyId:
          type: integer
          format: int64
          nullable: true
          description: 공구 ID. 생성 전 단계면 null
          example: 960004
        files:
          type: array
          description: 발급 대상 파일 목록
          items:
            $ref: '#/components/schemas/ImagePresignedUploadItemRequest'
    ImagePresignedUploadItemResponse:
      type: object
      required: [category, key, uploadUrl, method]
      properties:
        category:
          $ref: '#/components/schemas/ImageUploadCategory'
        key:
          type: string
          description: S3 Object Key
          example: dev/group-buys/pending/7/20260529/products/550e8400-e29b-41d4-a716-446655440000.jpg
        uploadUrl:
          type: string
          description: S3 업로드용 Presigned URL
          example: https://moongchijang-s3.s3.ap-northeast-2.amazonaws.com/dev/group-buys/960004/products/...
        method:
          type: string
          example: PUT
    ImagePresignedUploadResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          description: 발급 결과 목록
          items:
            $ref: '#/components/schemas/ImagePresignedUploadItemResponse'
    ApiResponseImagePresignedUpload:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/ImagePresignedUploadResponse'
        error: { nullable: true, example: null }
    ImageDeleteRequest:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          description: 삭제할 S3 key 목록
          items:
            type: string
          example:
            - dev/group-buys/960004/thumbnail/550e8400-e29b-41d4-a716-446655440000.jpg
            - dev/group-buys/960004/products/550e8400-e29b-41d4-a716-446655440001.jpg
    ImageDeleteResponse:
      type: object
      required: [deletedKeys, failedKeys]
      properties:
        deletedKeys:
          type: array
          description: 삭제 완료 key 목록
          items:
            type: string
        failedKeys:
          type: array
          description: 삭제 실패 key 목록
          items:
            type: string
    ApiResponseImageDelete:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/ImageDeleteResponse'
        error: { nullable: true, example: null }